+++ NEWS.md
@@ -1,3 +1,15 @@
+# ctv 0.9-2
+
+* Bug fix in `read.ctv()` when `pkg(..., priority = "core")` is used more than
+  once on the same package name (reported by Rocio Joo).
+  
+* Sorting the package list is done ignoring the case now.
+
+* Various improvements in `ctv2html()`: Include explicit citation in header table,
+  include instructions for contributions and installation (if `cran = TRUE`), more
+  compact layout of package list, handle task views without links section correctly.
+
+
 # ctv 0.9-1
 
 * Bug fix in `read.ctv(file)` when `file` is actually a full path and not just